Most recent version of "PHP INI Editor Class" published at this URL:
https://www.phpclasses.org/package/10165-PHP-View-and-edit-INI-files-in-Web-forms.html

Header is as follows:

Author: Blupixel IT Srl
Last Modifcation Date: 23 Jan 2017
Website: www.blupixelit.eu
support at: support@blupixelit.eu

Copyright (c) 2017 Blupixel IT Srl - Trento (Italy)

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <http://www.gnu.org/licenses/>.

This class provide an easy editor for ini files

Use it as described in the sample.php file or using the code below
Remeber to give write permission to the ini file and to the backup folder